requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
email-validator